Pasa la función constrain a float

+++ src/src/utils.c
@@ -4,7 +4,7 @@ float map(float x, float in_min, float in_max, float out_min, float out_max) {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
 
-int32_t constrain(int32_t x, int32_t min, int32_t max) {
+float constrain(float x, float min, float max) {
   if (x > max) {
     return max;
   } else if (x < min) {
